COMP: Fix -Wsign-compare, -Wunused-parameter and -Wunused-variable warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=27990
git-svn-id: http://svn.slicer.org/Slicer4/trunk@27990 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/Loadable/Segmentations/Logic/vtkSlicerSegmentationsModuleLogic.cxx
@@ -2345,7 +2345,7 @@ std::string vtkSlicerSegmentationsModuleLogic::GetSafeFileName(std::string origi
   // (same method as in qSlicerFileNameItemDelegate::fixupFileName)
   std::string safeName("");
   vtksys::RegularExpression regExp("[A-Za-z0-9\\ \\-\\_\\.\\(\\)\\$\\!\\~\\#\\'\\%\\^\\{\\}]");
-  for (int i=0; i<originalName.size(); ++i)
+  for (size_t i=0; i<originalName.size(); ++i)
     {
     std::string currentCharStr("");
     currentCharStr += originalName[i];
